Avoid italicizing entire subs
Some Amazon subtitles use <span> for italics, but some do not, and have a lot of <span> tags which are false positives. Span will only be treated as italics now, if there isn't are no span elements with an explicit italics value now.

Additionaly, a check has been added to remove italics if every single line has them, as this is almost guaranteed to be a mistake.

The italics sample also suffered from an issue with needless spaces being kept in after italics tags were merged into one, this has now been fixed.