Add a CLI `subby` script that can convert and process subs
This uses click for the CLI arguments and there-abouts, and poetry registers the `subby` application via scripts config.

Do note that when adding `click` dependency, this has updated the poetry lock format from v1.1 to v2.0. I also added `__version__` to the package constructor file so that I can reference the version with `subby -v`. This means when updating the version you must take the liberty of manually updating the version in the init py file and pyproject file.