(WebVTT) Parse line:0 as top position
This is technically in the specs, but I've only seen a sample with it for the first time. It seems to use -1 and -2 for regular lines, depending on whether it's a single or a double line, so it seems fine to special-case it

00:01:07.109 --> 00:01:12.114 align:middle line:0
BEM-VINDO A TWIN PEAKS
POPULAÇÃO: 51.201

00:03:13.944 --> 00:03:15.487 align:middle line:-1
Eu vou pescar.

00:03:40.971 --> 00:03:44.933 align:middle line:-2
E toca a solitária buzina
de nevoeiro.